description
In this demonstration, I will distinguish between classification techniques and their implicit assumptions. I will be creating basic classification models using statsmodels and sklearn then evaluate and interpret the performance of the outputs


The dataset gives information about the employees in the organization and the target variable (Y) is Attrition, which tells whether an employee was replaced or not.  We will look at the variables and different factors that affect employee performance, including factors like work from home sitaution, job role, monthly income etc.

The original dataset contains around 35 feature variables. I will be using a subset of that and use 14 features in this exercise